Add files via upload
The Online E-Shop is a modern and responsive e-commerce application built using React.js. It provides a seamless shopping experience for users, offering features like product browsing, searching, filtering, and a user-friendly cart and checkout system. The platform is designed to be scalable, customizable, and suitable for various product categories, making it ideal for small to medium-sized businesses.